[executor] extract storage integration test for code reuse
