+++ execution/executor-test-helpers/Cargo.toml
@@ -10,6 +10,8 @@ publish = false
 edition = "2018"
 
 [dependencies]
+anyhow = "1.0.31"
+rand = "0.7.3"
 tempfile = "3.1.0"
 
 config-builder = { path = "../../config/config-builder", version = "0.1.0" }
@@ -24,3 +26,4 @@ libra-vm = { path = "../../language/libra-vm", version = "0.1.0" }
 libra-workspace-hack = { path = "../../common/workspace-hack", version = "0.1.0" }
 storage-interface = { path = "../../storage/storage-interface", version = "0.1.0" }
 storage-service = { path = "../../storage/storage-service", version = "0.1.0" }
+transaction-builder = { path = "../../language/transaction-builder", version = "0.1.0" }
